Develop

+++ components/delete-spanish.tsx
@@ -10,10 +10,13 @@ type Props = {
 export default function DeleteSpanish({ type, id, callback }: Props) {
   const clickHandler = async () => {
     try {
-      await deleteSpanish(type, id);
+      const ok = confirm('삭제하시겠습니까?');
+      if (ok) {
+        await deleteSpanish(type, id);
 
-      if (callback) {
-        callback();
+        if (callback) {
+          callback();
+        }
       }
     } catch (err) {
       console.log(err);
@@ -21,8 +24,8 @@ export default function DeleteSpanish({ type, id, callback }: Props) {
   };
 
   return (
-    <span onClick={clickHandler}>
-      <IoMdRemoveCircle className='text-red-600' />
+    <span onClick={clickHandler} className='cursor-pointer'>
+      <IoMdRemoveCircle className='text-yellow-400' />
     </span>
   );
 }
